+++ tests/test_api_good_nonlocal.py
@@ -53,7 +53,7 @@ def skip_if_no_nonloc(pytestconfig):
     ids=(lambda x: "" if "://" in x else x),
 )
 @pytest.mark.timeout(30)
-def test_api_inventory_known_header_required(name, url, subtests):
+def test_api_inventory_known_header_required(name, url):
     """Confirm URL load works on docs pages requiring HTTP header config."""
     inv = soi.Inventory(url=url)
     assert inv.count > 0
@@ -68,7 +68,6 @@ def test_api_inventory_many_url_imports(
     misc_info,
     sphinx_load_test,
     pytestconfig,
-    subtests,
 ):
     """Confirm a plethora of .inv files downloads properly via url arg.
 
@@ -91,21 +90,10 @@ def test_api_inventory_many_url_imports(
     inv2 = soi.Inventory(url=misc_info.remote_url.format(proj_name))
 
     # Test the things
-    with subtests.test(msg="properties"):
-        assert inv1.project == inv2.project
-        assert inv1.version == inv2.version
-        assert inv1.count == inv2.count
-        for objs in zip(inv1.objects, inv2.objects):
-            assert objs[0].name == objs[1].name
-            assert objs[0].domain == objs[1].domain
-            assert objs[0].role == objs[1].role
-            assert objs[0].uri == objs[1].uri
-            assert objs[0].priority == objs[1].priority
-            assert objs[0].dispname == objs[1].dispname
+    assert inv1 == inv2
 
     # Ensure sphinx likes the regenerated inventory
-    with subtests.test(msg="sphinx_load"):
-        data = inv2.data_file()
-        cmp_data = soi.compress(data)
-        soi.writebytes(str(scr_fpath), cmp_data)
-        sphinx_load_test(scr_fpath)
+    data = inv2.data_file()
+    cmp_data = soi.compress(data)
+    soi.writebytes(scr_fpath, cmp_data)
+    sphinx_load_test(scr_fpath)